fixes after review

+++ scripts/observers/clean.py
@@ -1,12 +1,18 @@
 import shutil
 import os
+import sys
 from pathlib import Path
 from dotenv import load_dotenv
 
 
 def main():
     load_dotenv()
-    working_dir = Path(Path.home() / os.getenv('WORKING_DIRECTORY'))
+    working_dir_var = os.getenv('WORKING_DIRECTORY')
+    if working_dir_var == "":
+        print("working directory folder name cannot be empty")
+        sys.exit()
+
+    working_dir = Path(Path.home() / working_dir_var)
 
     try:
         shutil.rmtree(working_dir)
